requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scipy>=1.11
jinja2>=3.1
pyyaml>=6.0
click>=8.1
networkx>=3.1
seaborn>=0.12
tabulate>=0.9
